% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.formula.R
\name{cox.formula}
\alias{cox.formula}
\title{cox.formula
Generate a cox formula.}
\usage{
cox.formula(y, x, x.covars, frailty = NULL)
}
\arguments{
\item{y}{`character vector` with `length = 1`. Dependent variables including survival time and ending events.}

\item{x}{`character vector` with `length = 1`. Independent variables of interest.}

\item{x.covars}{`character vector`. Covariates other than the independent variable.}

\item{frailty}{`character vector` with `length = 1`. The default is NULL. you can set it if you are fitting a frailty cox model.}
}
\value{
a `formula`
}
\description{
cox.formula
Generate a cox formula.
}
