% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.result.R
\name{cox.result}
\alias{cox.result}
\title{cox.result}
\usage{
cox.result(fit)
}
\arguments{
\item{fit}{An object with class is `coxph` or `coxme`.}
}
\value{
A tibble.
}
\description{
Get the results of the independent variables of interest in the cox model
}
\details{
`x.vars`: character vector. Name of the independent variable of interest.
`term`: character vector. When the independent variable of interest is numeric, term and x.vars are consistent.
      When the independent variable of interest is a factor, x.vars will be merged with the factor level in term.
`HR.POS`: character vector. "positive" means HR >= 1; "negetive" means HR <1.
`p.sig`: character vector. "yes" means p.value <=0.05; "no" means p.sig > 0.05.
`estimate`: numeric vector. The coefficient in cox model of x.vars.
`std.error`: numeric vector. The standard error in cox model of x.vars.
`HR`: numeric vector. The hazard ratio in cox model of x.vars. HR = exp(estimate)
`HR.LOW`: numeric vector. Lower limit of 95% confidence interval.
`HR.UP`: numeric vector. Upper limit of 95% confidence interval.
`p.value`: numeric vector. P value of x.vars.
}
