% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.time.R
\name{cox.time}
\alias{cox.time}
\title{cox.time}
\usage{
cox.time(date, outcome)
}
\arguments{
\item{date}{date or numeric vector. The date of the follow-up visit, which can be year-month-day in date format. It can also be a year, month or day in separate numerical values.}

\item{outcome}{numeric vector. There are only two values, 0: no ending event has occurred; 1: an ending event has occurred.}
}
\value{
numeric vector with lengh = 1.
}
\description{
Calculating the survival time of patients
}
